ai-runtime/pyproject.toml
[tool.poetry]
name = "ai-runtime"
version = "0.1.0"
description = "AI Runtime Service with FastAPI and LangChain"
authors = ["Gerry Tao <gerry.tao@outlook.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "ai_runtime", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.13"
fastapi = ">=0.128.0,<0.129.0"
uvicorn = {extras = ["standard"], version = ">=0.40.0,<0.41.0"}
openai = "^1.59.12"
pymilvus = "^2.5.5"
weaviate-client = "^4.9.0"
langchain-text-splitters = "^0.3.4"
pydantic-settings = "^2.7.1"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-mock (>=3.15.1,<4.0.0)"
]
